backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0

# Static Files
whitenoise>=6.6,<7.0

# Excel Processing
pandas>=2.1,<3.0
openpyxl>=3.1,<4.0
xlrd>=2.0,<3.0

# Production Server
gunicorn>=21.0,<23.0
